--- a/.gitignore
+++ b/.gitignore
@@ -163,3 +163,4 @@ cython_debug/
 
 .vscode
 vcz_test_cache/
+**/.DS_Store