--- a/performance/data/Makefile
+++ b/performance/data/Makefile
@@ -16,14 +16,14 @@ all: sim_10k.vcz
 sim_10k.ts:
 	stdpopsim HomSap -c chr22 -o sim_10k.ts pop_0:10000
 
-sim_10k.vcf.gz: sim_10k.ts
-	tskit vcf sim_10k.ts | bgzip > sim_10k.vcf.gz
+%.vcf.gz: %.ts
+	tskit vcf $< | bgzip > $@
 
-sim_10k.vcf.gz.csi: sim_10k.vcf.gz
-	bcftools index sim_10k.vcf.gz
+%.vcf.gz.csi: %.vcf.gz
+	bcftools index $<
 
-sim_10k.vcz: sim_10k.vcf.gz sim_10k.vcf.gz.csi
-	vcf2zarr convert sim_10k.vcf.gz sim_10k.vcz
+%.vcz: %.vcf.gz %.vcf.gz.csi
+	vcf2zarr convert $< $@
 
 clean:
 	rm -rf sim_10k.*