@jeromekelleher 2024-10-17
Excellent! Can you profile the slower commands please so we can figure out what needs to be done?

I would be good to get a baseline "time to output all fields" on the real data as well, so we can get a feel for how well the C VCF text generating is doing.